refactor: `MyBatis` `DibRepository` 리팩터
- 파일 이름 변경 `(DibRepositoryImpl -> MyBatisDibRepository)`
- 파일 위치 변경
- `saveNewDib` 메서드 `Optional -> DibEntity` 로 반환 타입 변경
