+++ application/src/main/java/core/application/users/repositories/DibRepository.java
@@ -1,10 +1,7 @@
 package core.application.users.repositories;
 
-import core.application.users.models.entities.DibEntity;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
+import core.application.users.models.entities.*;
+import java.util.*;
 
 /**
  * {@code DIB_TABLE} 과 관련된 {@code Repository}
@@ -16,12 +13,11 @@ public interface DibRepository {
     /**
      * 주어진 유저가 소유한 새로운 찜 목록을 DB 에 등록
      *
-     * @param userId 찜 목록을 소유할 유저 ID
+     * @param userId  찜 목록을 소유할 유저 ID
      * @param movieId 찜 목록에 들어갈 영화 ID
      * @return @return {@link DibEntity} 등록된 찜 정보
      */
-    Optional<DibEntity> saveNewDib(UUID userId, String movieId);
-
+    DibEntity saveNewDib(UUID userId, String movieId);
 
     //<editor-fold desc="READ">
 
@@ -34,10 +30,25 @@ public interface DibRepository {
     Optional<DibEntity> findByDibId(Long id);
 
 
+    /**
+     * 특정 유저의 찜 목록 불러오기
+     *
+     * @param userId 유저 ID
+     * @return 유저가 저장한 찜 목록들
+     */
     List<DibEntity> findByUserId(UUID userId);
 
+    /**
+     * 특정 유저의 특정 영화 찜 목록 가져오기 {@code (???)}
+     * <p>
+     * 무슨 역할인지 모르겠음...
+     *
+     * @param userId  유저 ID
+     * @param movieId 영화 ID
+     * @return 찜 목록
+     */
     Optional<DibEntity> findByUserIdAndMovieId(UUID userId, String movieId);
-  
+
     /**
      * 특정 영화의 찜 된 횟수를 반환
      *
@@ -53,7 +64,6 @@ public interface DibRepository {
      */
     List<DibEntity> selectAll();
     //</editor-fold>
-
 
     //<editor-fold desc="DELETE">
 
@@ -71,6 +81,12 @@ public interface DibRepository {
      */
     void deleteDib(UUID userId);
 
+    /**
+     * 특정 유저의 특정 찜 목록을 삭제
+     *
+     * @param userId  유저 ID
+     * @param movieId 삭제할 영화 ID
+     */
     void deleteDib(UUID userId, String movieId);
     //</editor-fold>
 }
